Refactoring
'data' is a better name for this module than 'models'
